+++ wheat-modern/src/main/scala/x7c1/wheat/modern/database/selector/presets/ClosableSequenceLoader.scala
@@ -8,9 +8,10 @@ import x7c1.wheat.modern.database.selector.SelectorProvidable.Implicits.Selector
 import x7c1.wheat.modern.database.selector.presets.ClosableSequenceLoader.{LoadingDone, LoadingError, SqlError}
 import x7c1.wheat.modern.database.selector.{CanIdentify, CanProvideSelector}
 import x7c1.wheat.modern.decorator.Imports.toRichView
+import x7c1.wheat.modern.fate.FutureFate
 import x7c1.wheat.modern.formatter.ThrowableFormatter.format
-import x7c1.wheat.modern.kinds.FutureFate.HasContext
-import x7c1.wheat.modern.kinds.{Fate, FutureFate}
+import x7c1.wheat.modern.fate.FutureFate.HasContext
+import x7c1.wheat.modern.kinds.Fate
 import x7c1.wheat.modern.sequence.Sequence
 
 import scala.language.higherKinds
