+++ wheat-modern/src/main/scala/x7c1/wheat/modern/observer/recycler/order/OrderUpdater.scala
@@ -7,7 +7,7 @@ import x7c1.wheat.modern.database.selector.presets.ClosableSequenceLoader
 import x7c1.wheat.modern.database.selector.presets.ClosableSequenceLoader.LoadingError
 import x7c1.wheat.modern.decorator.Imports.toRichView
 import x7c1.wheat.modern.kinds.Fate
-import x7c1.wheat.modern.kinds.FutureFate.HasContext
+import x7c1.wheat.modern.fate.FutureFate.HasContext
 import x7c1.wheat.modern.observer.recycler.order.DraggableSequenceRoute.{DragFinished, DragStarted, OnDragListener}
 
 import scala.language.{higherKinds, reflectiveCalls}
